On my computer for local testing, this takes the total test time from ~40 seconds to ~32 seconds.

The analysis folder tests go from ~9s to ~4s.